Add type hints as stub files
Stub files were initially generated using stubgen tool (from mypy), then
completed based on signatures and documentation and finally adjusted to
achieve 'mypy --strict blessed/' success and limit errors with 'mypy
--check-untyped-defs tests/'.

Type annotation for formatter's object's __call__() method is a bit
special. We make sure that all of them declares accepting arguments of
type Union[int, str] though some of them only actually accepts either
int or str argument. This is so that the return value of
Terminal.__getattr__() is an object with a consistent interface. This is
implemented using overloads and we use the special NoReturn to indicate
invalid input type.